Read ~/.memorc for additional configuration.
It's now possible to create a configuration file to set an alternative
path where Memo saves the notes. ~/.memorc is now used for the
configuration and an alternative path can be set using property
MEMO_PATH=/some/path/here.

This commit also refactors some of the functions in memo.c to make them
more general and more reusable.

Signed-off-by: Niko Rosvall <niko@ideabyte.net>